- added custom boss death actions.
These work roughly like MBF's A_LineEffect, but are a bit more thorough about what is allowed and avoid hacking the first player to trigger the action.
For the definition it uses the same class names as ZDoom because that is the closest to a naming standard there is.

+++ prboom2/src/p_setup.c
@@ -221,7 +221,7 @@ static dboolean CheckForIdentifier(int lumpnum, const byte *id, size_t length)
 {
   dboolean result = false;
 
-  if (W_LumpLength(lumpnum) >= length)
+  if (W_LumpLength(lumpnum) >= (int)length)
   {
     const char *data = W_CacheLumpNum(lumpnum);
